Added rule to ignore grant routes on lout documentation
When using the plugin `lout` with `hapi`, the routes generated with
`grant` appear in the documentation generated. As these are private
routes, I don’t believe they should be shown.

This fix adds the ignore rules as specified in the `lout` documentation:
https://github.com/hapijs/lout#ignoring-a-route-in-documentation

Without the plugin installed the configuration is ignored.